+++ packages/frontend/src/components/dialogs/CreateChat/index.tsx
@@ -62,6 +62,7 @@ import ViewProfile from '../ViewProfile'
 import { isInviteLink } from '@deltachat-desktop/shared/util'
 import { copyToBlobDir } from '../../../utils/copyToBlobDir'
 import { useRpcFetch } from '../../../hooks/useFetch'
+import { I18nContext } from '../../../contexts/I18nContext'
 
 const enum GroupType {
   /**
@@ -90,7 +91,7 @@ export default function CreateChat(props: DialogProps) {
   const [viewMode, setViewMode] = useState<ViewMode>('main_')
 
   return (
-    <Dialog width={400} onClose={onClose} fixed>
+    <Dialog width={400} onClose={onClose} fixed dataTestid='create-chat-dialog'>
       {viewMode == 'main_' && <CreateChatMain {...{ setViewMode, onClose }} />}
       {viewMode == GroupType.REGULAR_GROUP && (
         <>
@@ -161,7 +162,7 @@ type CreateChatMainProps = {
 
 function CreateChatMain(props: CreateChatMainProps) {
   const { setViewMode, onClose } = props
-  const tx = useTranslationFunction()
+  const { tx, writingDirection } = useContext(I18nContext)
   const openConfirmationDialog = useConfirmationDialog()
   const accountId = selectedAccountId()
   const { openDialog } = useDialog()
@@ -199,7 +200,16 @@ function CreateChatMain(props: CreateChatMainProps) {
   // Chatmail accounts can't send unencrypted emails. See
   // - https://github.com/deltachat/deltachat-desktop/issues/5294#issuecomment-3089552788
   // - https://github.com/deltachat/deltachat-ios/blob/a0043be425d9c14f4039561957adb82ef1ab2adb/deltachat-ios/Controller/NewChatViewController.swift#L76-L78
-  const showNewEmail = !isChatmail && queryStr.length === 0
+  const [forceShowNewEmail_, setForceShowNewEmail_] = useState(false)
+  useEffect(() => {
+    // For E2E tests.
+    window.__testForceShowNewEmailButton = setForceShowNewEmail_
+    return () => {
+      window.__testForceShowNewEmailButton = undefined
+    }
+  }, [])
+  const showNewEmail =
+    (!isChatmail && queryStr.length === 0) || forceShowNewEmail_
 
   const showAddContact = !(
     isChatmail ||
@@ -399,6 +409,7 @@ function CreateChatMain(props: CreateChatMainProps) {
                     outerRef={fixedSizeListOuterRef}
                     height={height}
                     width='100%'
+                    direction={writingDirection}
                     // TODO fix: The size of each item is determined
                     // by `--local-avatar-size` and `--local-avatar-vertical-margin`,
                     // which might be different, e.g. currently they're smaller for