+++ packages/frontend/src/hooks/chat/useCreateChatByEmail.ts
@@ -31,6 +31,12 @@ export default function useCreateChatByEmail(): CreateChatByEmail {
       if (chatId) {
         return chatId
       }
+      /**
+       * Although we know that creating a new chat will probably fail on chatmail accounts,
+       * since we can't create an encrypted chat based on an email address, we continue
+       * here.
+       * see https://github.com/deltachat/deltachat-android/issues/3361
+       */
 
       // Ask user if they want to proceed with creating a new contact and / or chat
       const continueProcess = await new Promise((resolve, _reject) => {